Add REST endpoint for GETting messages
This commit adds two endpoints:
/messages returns all the messages from ES, in a paginated way. Default
page size is 10 and the default page is 0, both can be overridden by
query parameters. The result contains the total hit count and the hits
of the current page, in escaped form.
/messages/{position} returns one given message from ES, unescaped.

Messages are ordered by input timestamp, that ordering is used in both
cases.